users

+++ fuelfinder/users/views.py
@@ -28,12 +28,15 @@ def suppliers_delete(request, sid):
 
 def buyers_list(request):
     buyers = BuyerProfile.objects.all()
-    return render(request, 'users/buyers_list.html', {'buyers': buyers})
+    edit_form = BuyerProfileEditForm()
+    delete_form = ActionForm()
+    return render(request, 'users/buyers_list.html', {'buyers': buyers, 'edit_form': edit_form, 'delete_form': delete_form})
 
-def delete(request, phone_number):
-    #byr = get_object_or_404(BuyerProfile, phone_number)
-    buyer = BuyerProfile.objects.filter(phone_number=phone_number).first()
-    buyer.delete()
+def buyers_delete(request, sid):
+    buyer = BuyerProfile.objects.filter(id=sid).first()
+    if request.method == 'POST':
+        buyer.delete()    
+
     return redirect('users:buyers_list')
 
 # Begining Of Supplier Management
@@ -86,6 +89,53 @@ def supplier_user_create(request, sid):
     return render (request, 'users/add_user.html', {'form': form, 'supplier': supplier, 'staff': staff, 'count': count,
      'delete_form':delete_form, 'edit_form': edit_form, 'profile_form':profile_form}) 
 
+
+def buyer_user_create(request, sid):
+    buyer = get_object_or_404(BuyerProfile, id=sid) 
+    staff = BuyerContact.objects.filter(buyer_profile=buyer)
+    count = BuyerContact.objects.all().count()
+    delete_form = ''
+    edit_form = ''
+    if request.method == 'POST':
+        user_count = BuyerContact.objects.filter(buyer_profile=buyer).count()
+        if user_count > 10:
+            raise Http404("Your organisation has reached the maximum number of users, delete some ")
+        form = BuyerContactForm(request.POST)
+        profile_form = UserUpdateForm(request.POST, instance=buyer)
+
+        if profile_form.is_valid():
+            buyer = profile_form.save()
+            messages.success(request, 'Your Changes Have Been Saved')
+            return redirect('users:buyer_user_create', sid=buyer.id)
+
+        
+
+        if form.is_valid():
+            email = form.cleaned_data['email']
+            password = form.cleaned_data['password']
+            cellphone = form.cleaned_data['phone']
+            user = User.objects.create_user(email, email, password)
+            user.last_name = form.cleaned_data['last_name']
+            user.first_name = form.cleaned_data['first_name']
+            user.save()   
+            contact = BuyerContact.objects.create(user=user, phone=cellphone, buyer_profile=buyer)
+            #contact.save()
+            messages.success(request, ('Your profile was successfully updated!'))
+            return redirect('users:buyer_user_create', sid=buyer.id)
+            
+        
+        else:
+            msg = "Error in Information Submitted"
+            messages.error(request, msg)
+    else:
+        form = BuyerContactForm()
+        profile_form = UserUpdateForm(instance=buyer)
+
+
+
+    return render (request, 'users/add_buyer.html', {'form': form, 'buyer': buyer, 'staff': staff, 'count': count, 'delete_form':delete_form, 'edit_form': edit_form, 'profile_form':profile_form}) 
+
+
 def edit_supplier(request,id):
     supplier = get_object_or_404(SupplierProfile, id=id)
     if request.method == "POST":
@@ -93,11 +143,24 @@ def edit_supplier(request,id):
         if form.is_valid():
             data = form.cleaned_data
             supplier = form.save()
-            messages.success(request, 'Changes Successfully Update')
+            messages.success(request, 'Changes Successfully Updated')
             return redirect('users.index')
     else:
         form = SupplierProfile(instance=supplier)
     return render(request, 'users/supplier_edit.html', {'form': form, 'supplier': supplier})
+
+def edit_buyer(request,id):
+    buyer = get_object_or_404(BuyerProfile, id=id)
+    if request.method == "POST":
+        form = UserForm(request.POST, request.FILES, instance=buyer)
+        if form.is_valid():
+            data = form.cleaned_data
+            buyer = form.save()
+            messages.success(request, 'Changes Successfully Updated')
+            return redirect('users.index')
+    else:
+        form = BuyerProfile(instance=supplier)
+    return render(request, 'users/buyer_edit.html', {'form': form, 'buyer': buyer})
 
 def delete_user(request,id):
     supplier = get_object_or_404(SupplierProfile, id=id)
